Simplify grammar around binary operator modifiers
The Prometheus PromQL parser inserts an intermediate 1:1 rule that we
copied initially, but that we probably don't need:

https://github.com/prometheus/prometheus/blob/511511324adfc4f4178f064cc104c2deac3335de/promql/parser/generated_parser.y#L247-L249

Also, defining BoolModifier as either an empty string or "bool" caused
lezer to put a BoolModifier token in places where it shouldn't be
grammatically allowed, so now just defined it as an optional "bool"
string instead.

Finally, I ended up condensing a bunch of rules into one.

Signed-off-by: Julius Volz <julius.volz@gmail.com>